Start work on Event type

+++ sdl-sys/src/lib.rs
@@ -2,22 +2,24 @@
 #![allow(non_camel_case_types)]
 #![allow(non_snake_case)]
 #![allow(improper_ctypes)]
+#![allow(deref_nullptr)]
 #![no_std]
 
 include!(concat!(env!("OUT_DIR"), "/sdl_bindings.rs"));
 
+#[repr(u32)]
 pub enum SDL_WindowFlags {
-    SDL_SWSURFACE = crate::SDL_SWSURFACE as isize,
-    SDL_HWSURFACE = crate::SDL_HWSURFACE as isize,
-    SDL_ASYNCBLIT = crate::SDL_ASYNCBLIT as isize,
-    SDL_ANYFORMAT = crate::SDL_ANYFORMAT as isize,
-    SDL_HWPALETTE = crate::SDL_HWPALETTE as isize,
-    SDL_DOUBLEBUF = crate::SDL_DOUBLEBUF as isize,
-    SDL_FULLSCREEN = crate::SDL_FULLSCREEN as isize,
-    SDL_OPENGL = crate::SDL_OPENGL as isize,
-    SDL_OPENGLBLIT = crate::SDL_OPENGLBLIT as isize,
-    SDL_RESIZABLE = crate::SDL_RESIZABLE as isize,
-    SDL_NOFRAME = crate::SDL_NOFRAME as isize,
+    SDL_SWSURFACE = crate::SDL_SWSURFACE,
+    SDL_HWSURFACE = crate::SDL_HWSURFACE,
+    SDL_ASYNCBLIT = crate::SDL_ASYNCBLIT,
+    SDL_ANYFORMAT = crate::SDL_ANYFORMAT,
+    SDL_HWPALETTE = crate::SDL_HWPALETTE,
+    SDL_DOUBLEBUF = crate::SDL_DOUBLEBUF,
+    SDL_FULLSCREEN = crate::SDL_FULLSCREEN,
+    SDL_OPENGL = crate::SDL_OPENGL,
+    SDL_OPENGLBLIT = crate::SDL_OPENGLBLIT,
+    SDL_RESIZABLE = crate::SDL_RESIZABLE,
+    SDL_NOFRAME = crate::SDL_NOFRAME,
 }
 
 #[cfg(feature = "mixer")]
